pyproject.toml
[project]
name = "document-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.4",
    "langchain>=0.3.18",
    "langchain-community>=0.3.17",
    "langchain-google-genai>=2.0.9",
    "pypdf>=5.3.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
]
